add onekey fee
