+++ packages/kit/src/views/Swap/pages/components/SwapQuoteResult.tsx
@@ -1,9 +1,9 @@
-import { memo, useCallback, useRef, useState } from 'react';
+import { memo, useCallback, useMemo, useRef, useState } from 'react';
 
 import BigNumber from 'bignumber.js';
 import { useIntl } from 'react-intl';
 
-import type { IDialogInstance } from '@onekeyhq/components';
+import type { IDialogInstance, IImageSourceProps } from '@onekeyhq/components';
 import {
   Accordion,
   Dialog,
@@ -11,7 +11,10 @@ import {
   Icon,
   NumberSizeableText,
   SizableText,
+  Stack,
+  Image,
   XStack,
+  YStack,
 } from '@onekeyhq/components';
 import { useDebounce } from '@onekeyhq/kit/src/hooks/useDebounce';
 import {
@@ -44,6 +47,14 @@ import { useSwapQuoteLoading } from '../../hooks/useSwapState';
 import SwapApproveAllowanceSelectContainer from './SwapApproveAllowanceSelectContainer';
 import SwapSlippageContentContainer from './SwapSlippageContentContainer';
 import SwapSlippageTriggerContainer from './SwapSlippageTriggerContainer';
+
+interface IProtocolFeeInfo {
+  name: string;
+  fee: number;
+  color: string;
+  icon: IImageSourceProps['source'];
+  maxFee: number;
+}
 
 interface ISwapQuoteResultProps {
   quoteResult?: IFetchQuoteResult;
@@ -183,6 +194,60 @@ const SwapQuoteResult = ({
     slippageOnSave,
     setSwapSlippageDialogOpening,
   ]);
+  const protocolFeeInfoList: IProtocolFeeInfo[] = useMemo(
+    () => [
+      {
+        maxFee: 0.875,
+        name: 'metamask',
+        icon: {
+          uri: 'https://uni.onekey-asset.com/static/logo/metamasklogo.png',
+        },
+        fee: 0.875,
+        color: '#F5841F',
+        url: 'https://uni.onekey-asset.com/static/logo/metamasklogo.png',
+      },
+      {
+        maxFee: 0.875,
+        name: 'zerion',
+        fee: 0.8,
+        color: '#2461ED',
+
+        icon: {
+          uri: 'https://uni.onekey-asset.com/static/logo/zerionlogo.png',
+        },
+      },
+      {
+        maxFee: 0.875,
+        name: 'oneKey',
+        fee: 0.3,
+        // color: '#202020',
+        color: '$bgInverse',
+        icon: require('@onekeyhq/kit/assets/logo.png'),
+      },
+    ],
+    [],
+  );
+  const renderProtocolFeeListItem = useCallback(
+    (item: IProtocolFeeInfo) => (
+      <XStack gap="$3" alignItems="center">
+        <Stack w={20} h={20}>
+          <Image source={item.icon} w={16} h={16} />
+        </Stack>
+        <Stack flex={1}>
+          <Stack
+            backgroundColor={item.color}
+            borderRadius="$full"
+            width={`${(item.fee / item.maxFee) * 100}%`}
+            height="$1"
+          />
+        </Stack>
+        <SizableText size="$bodySm" color="$text" textAlign="right">
+          {item.fee}%
+        </SizableText>
+      </XStack>
+    ),
+    [],
+  );
   const fromAmountDebounce = useDebounce(fromTokenAmount, 500, {
     leading: true,
   });
@@ -310,6 +375,35 @@ const SwapQuoteResult = ({
                     </NumberSizeableText>
                   }
                 />
+              ) : null}
+              {quoteResult?.fee?.percentageFee ? (
+                
+                <SwapCommonInfoItem
+                  title="OneKey fee"
+                  isLoading={swapQuoteLoading}
+                  valueComponent={
+                    <NumberSizeableText size="$bodyMdMedium" formatter="value">
+                      {`${quoteResult?.fee?.percentageFee}%`}
+                    </NumberSizeableText>
+                  }
+                  onPress={() => {
+                    Dialog.show({
+                      icon: 'OnekeyBrand',
+                      title: 'OneKey fee',
+                      description: intl.formatMessage({
+                        id: ETranslations.provider_ios_popover_onekey_fee_content,
+                      }),
+                      showCancelButton: false,
+                      renderContent: (
+                        <YStack>
+                          {protocolFeeInfoList.map((item) =>
+                            renderProtocolFeeListItem(item),
+                          )}
+                        </YStack>
+                      ),
+                    });
+                  }}
+                ></SwapCommonInfoItem>
               ) : null}
               {swapTokenMetadata?.swapTokenMetadata
                 ? tokenMetadataParse(
